+++ judgels-backends/judgels-grader-app/src/main/java/judgels/JudgelsGraderApplication.java
@@ -2,8 +2,8 @@ package judgels;
 
 import io.dropwizard.core.Application;
 import io.dropwizard.core.setup.Environment;
-import judgels.grading.CacheModule;
 import judgels.grading.GradingModule;
+import judgels.grading.cache.CacheModule;
 import judgels.isolate.IsolateModule;
 import judgels.messaging.rabbitmq.RabbitMQModule;
