+++ judgels-backends/judgels-grader-app/src/main/java/judgels/JudgelsGraderComponent.java
@@ -2,9 +2,9 @@ package judgels;
 
 import dagger.Component;
 import jakarta.inject.Singleton;
-import judgels.grading.CacheModule;
 import judgels.grading.GradingModule;
 import judgels.grading.GradingRequestPoller;
+import judgels.grading.cache.CacheModule;
 import judgels.isolate.IsolateModule;
 import judgels.messaging.rabbitmq.RabbitMQModule;
 import judgels.service.JudgelsModule;
