+++ judgels-backends/judgels-grader-app/src/main/java/judgels/grading/cache/CacheConfiguration.java
@@ -1,4 +1,4 @@
-package judgels.grading;
+package judgels.grading.cache;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import java.nio.file.Path;
